web_app_Streamlit/requirements.txt
PyMuPDF==1.24.5
openai==1.31.1
instructor==1.3.2
python-dotenv==1.0.1
streamlit==1.37.1

